LayoutTests/imported/w3c/web-platform-tests/trusted-types/block-string-assignment-to-Range-createContextualFragment-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


FAIL range.createContextualFragment assigned via policy (successful HTML transformation). undefined is not an object (evaluating 'win.trustedTypes.createPolicy')
FAIL `range.createContextualFragment(string)` throws. assert_throws_js: function "_ => {
      var result = range.createContextualFragment("A string");
    }" did not throw
FAIL `range.createContextualFragment(null)` throws. assert_throws_js: function "_ => {
      var result = range.createContextualFragment(null);
    }" did not throw
FAIL `range.createContextualFragment(string)` assigned via default policy (successful HTML transformation). undefined is not an object (evaluating 'window.trustedTypes.createPolicy')
PASS `range.createContextualFragment(null)` assigned via default policy does not throw.
